Documentacion, codigo mas limpio y manejo de excepciones

+++ abmc/src/main/java/com/abmc/abmc/Service/UserService.java
@@ -1,8 +1,11 @@
 package com.abmc.abmc.Service;
 
+import com.abmc.abmc.Exceptions.ResourceNotFoundException;
 import com.abmc.abmc.Persistence.UserMapper;
 import com.abmc.abmc.entities.Users;
+import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,13 +19,27 @@ public class UserService {
         return userMapper.findAll();
     }
     public Users findById(int id){
-        return userMapper.findById(id);
+        Users user = userMapper.findById(id);
+        if (user == null) {
+            throw new ResourceNotFoundException("User not found with ID: " + id);
+        }
+        return user;
     }
     public void deleteById(int id){
+        Users user = userMapper.findById(id);
+        if (user == null) {
+            throw new ResourceNotFoundException("Usuario no encontrado con ID: " + id);
+        }
         userMapper.deleteById(id);
     }
-    public void insert(Users user){
+
+    public ResponseEntity<String> insert(@Valid Users user){
+        Users existingUser = userMapper.findByEmail(user.getEmail());
+        if (existingUser != null) {
+            throw new IllegalArgumentException("Email is already in use: " + user.getEmail());
+        }
         userMapper.insert(user);
+        return ResponseEntity.ok("User inserted correctly");
     }
 
 }
